Fixed Update status by admin and rename file transaction
