+++ routes/transactionslips.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const {getTransactionSlips, getTransactionSlip, addTransactionSlip} = require("../controllers/transactionslip");
+const {getTransactionSlips, getTransactionSlip, addTransactionSlip} = require("../controllers/transactionslips");
 
 const router = express.Router({ mergeParams: true });
 const { protect, authorize } = require("../middleware/auth");
